macrursors: Improve performance
`macrursors--mark-all-instances-of`, `macrursors--mark-next-instance-of`, `macrursors--mark-previous-instance-of`:
Rewrite recursive calls as simple loops: elisp doesn't have tail-call
optimization so avoid filling the stack.

`macrursors--inside-secondary-selection`: Simplify test criterion.  This was an expensive function.  We don't need to look at all the overlays in the buffer, which `secondary-selection-exist-p` does.

`macrursors--mark-all-instances-of`, `macrursors--mark-next-instance-of`,
`macrursors--mark-previous-instance-of`: Since we are searching for a string,
not regexp, we don't need `re-search-forward`, replace with the simpler and much
faster `search-forward`.

`macrursors-mark-all-instances-of`,
`macrursors-mark-next-instance-of`,`macrursors-mark-previous-instance-of`: Use
`region-beginning`, `region-end` and `use-region-p` (where appropriate)

`macrursors--apply-kmacros`, `macrursors-apply-command`,
`macrursors--apply-command`: Use `execute-kbd-macro` instead of
`kmacro-call-macro`. The latter is meant for interactive use and does a bunch of
processing we don't need, like messaging and setting up a transient keymap.