macrursors-select: Better secondary selection
* macrursors-select.el (macrursors--select-set,
macrursors-select-types, macrursors-sel--filter-cursors,
macrursors-select--expand-cursors, macrursors-select,
macrursors-select--region, macrursors-select--types,
macrursors-select-map, macrursors-select-clear): It is now
possible to interact with the secondary selection using the
keyboard. Calling `macrursors-select' will cycle through secondary
selections covering objects listed in `macrursors-select-types',
which see. This command activates a transient keymap so it can be
called again with a single key. When doing so with
`macrursors-mode' active, macro cursors will be updated to the new
secondary selection.

* macrursors.el (macrursors-mark-all-lines,
macrursors--defun-mark-all, macrursors-mark-previous-instance-of,
macrursors-mark-next-instance-of,
macrursors--mark-all-instances-of): Store the current search
string or syntactic unit (word/sexp/line etc) in a new variable
so we can re-run the search when the secondary selection changes.
Note that the search only re-runs when the selection grows in
either direction around point. When the selection shrinks we
simply discard the cursors outside it.

(macrursors--instance): Buffer-local variable holding the last
thing we searched for to create cursors. (This is required to
update cursors when the secondary selection changes, but in the
future we can also use it to provide a macrursors-resume command
that reruns the last search.)