macrursors: make composable commands, add prefix args
macrursors.el (macrursors-mark-all-instances-of,
macrursors-mark-next-instance-of,
macrursors--mark-previous-instance-of,
macrursors-mark-next-instance-of,
macrursors-mark-previous-instance-of):
- Add numeric prefix args to -mark-next and -mark-previous commands
to mark multiple units at once.
- A negative prefix arg will remove the latest arg cursors.
- `macrursors-mark-next-instance-of' and
`macrursors-mark-previous-instance-of' now mark the next line if
there is no symbol at point and the next instance of the symbol at
point if the region is inactive.
- New commands `macrursors-mark-next-from-isearch' and
`macrursors-mark-previous-from-isearch' to mark the next and
previous instance of the current isearch string (including
regexps). Numeric prefix and negative args are supported.
- New commands `macrursors-mark-next-line',
`macrursors-mark-previous-line' to mark the next and previous
lines. Numeric prefix and negative args are supported.
- Composability pass: Every command can be followed by every
other. For example,
  + `-mark-all-instances-of` can be followed by `-mark-next` with
  a negative prefix arg to remove cursors,
  + `-mark-next-from-isearch` can be followed by `-mark-previous`
  to match more previous instances, and so on.

New functions:
macrursors--instance-with-bounds,
macrursors-mark-next-from-isearch,
macrursors-mark-previous-from-isearch, macrursors-mark-next-line,
macrursors-mark-previous-line